\name{duplicatedDataExample}
\alias{duplicatedDataExample}
\docType{data}
\title{
Tracking data example with duplicated timestamps
}
\description{
This file contains a data frame with fictional tracking data of two individuals, including duplicated timestamps. These data are used in the example of the function \code{\link{getDuplicatedTimestamps}}.
}

\examples{
data(duplicatedDataExample)
}
\keyword{datasets}
