\name{leroydbgb}
\alias{leroydbgb}
\docType{data}
\title{
dynamic Bivariate Gausian Bridge example object
}
\description{
Utilization densities calculated with dynBGB to exemplify functions.
}
\usage{data("leroydbgb")}

\details{
see createRDataFile.R in inst/extdata for the exact calculation
}

\examples{
data(leroydbgb)
leroydbgb
}
\keyword{datasets}
